Bug 34671: Ignore validation if patron attribute field is hidden
Patron attributes not used by patron category are still present in
form as hidden and marked as required if they are set as
mandatory. This prevents saving patron when validation is
required also for hidden fields. This patch adds and removes class "ignore_validation" for these fields based on their visibility.

To test:
1. Add patron attribute for patron category A and set it as mandatory.
2. Add patron with category B and fill in required fields.
3. Attempt to save.
=> Error is displayed above "Additional attributes and identifiers"
field set.
4. Apply this patch.
5. Repeat steps 2 and 3.
=> Save successfully.
6. Add new patron with category B, but this time change patrons
category as A from patron form.
=> Confirm that madatory patron attribute is displayed correctly
and without it patron cannot be saved.

Sponsored-by: Koha-Suomi Oy
Signed-off-by: Kelly <kelly@bywatersolutions.com>