Bug 36135: Add new tool to batch modify holds
We sometimes need to modify holds due library closing down etc.
For this we need tool to batch modify holds which allows us
to filter holds and modify them.

Tool allows users to select holds with following parameters:
 - expiration date
    -> if patron_expiration_date is set, it is used as main search filter
    -> otherwise expirationdate is used
 - library
 - found status
 - suspended status
 - suspend until date
 - hold note

Users are able to modify following  with tool:
 - expiration date (not expiration date set by patron)
 - pick up library
 - suspend status (only if hold has no found status)
 - suspend until date (only if suspend status is set)
 - hold note

To test:
1. Apply this patch.
2. Navigate to new tool Tools->Batch modify holds.
3. Fill in some criteria to search holds with.
=> Confirm that holds found match your search criteria.
4. Set e.g. new expiration date for all or just selected holds.
=> Confirm that holds were modified as they should have.
=> Confirm this also from database.

To test suspending holds with found status:
1. Find holds with found status.
2. Attempt to modify their suspend status.
=> Alert box with text "One or more holds have found status
and can't be suspended." should be displayed.

Play around, test different search compinations and modifications.